Upated allot of stuff

+++ syncly/models/base.py
@@ -1,5 +1,6 @@
 from diffsync import DiffSyncModel
 
+
 class Category(DiffSyncModel):
     """
     DiffSync model for a category.
@@ -7,11 +8,13 @@ class Category(DiffSyncModel):
     Attributes:
         name (str): The category name.
     """
+
     _modelname = "category"
     _identifiers = ("name",)
     _attributes = ()
 
     name: str
+
 
 class Package(DiffSyncModel):
     """
@@ -20,6 +23,7 @@ class Package(DiffSyncModel):
     Attributes:
         name (str): The package name.
     """
+
     _modelname = "package"
     _identifiers = ("name",)
     _attributes = ()
@@ -34,12 +38,12 @@ class Brand(DiffSyncModel):
     Attributes:
         name (str): The Brand name.
     """
+
     _modelname = "brand"
     _identifiers = ("name",)
     _attributes = ()
 
     name: str
-
 
 
 class Supplier(DiffSyncModel):
@@ -49,11 +53,13 @@ class Supplier(DiffSyncModel):
     Attributes:
         name (str): The supplier name.
     """
+
     _modelname = "supplier"
     _identifiers = ("name",)
     _attributes = ()
 
     name: str
+
 
 class Attribute(DiffSyncModel):
     """
@@ -63,13 +69,15 @@ class Attribute(DiffSyncModel):
         name (str): The attribute name.
         attribute_values (list): List of values for this attribute.
     """
+
     _modelname = "attribute"
-    _identifiers = ("name", )
+    _identifiers = ("name",)
     _attributes = ()
     _children = {"attribute_value": "attribute_values"}
 
     name: str
     attribute_values: list = []
+
 
 class AttributeValue(DiffSyncModel):
     """
@@ -79,12 +87,14 @@ class AttributeValue(DiffSyncModel):
         attribute (str): Identifier of the parent attribute.
         value (str): The value string.
     """
+
     _modelname = "attribute_value"
     _identifiers = ("attribute", "value")
     _attributes = ()
 
     attribute: str
     value: str
+
 
 class Product(DiffSyncModel):
     """
@@ -101,6 +111,7 @@ class Product(DiffSyncModel):
         attributes (list): Associated attribute values.
         photos (list): Associated product photos.
     """
+
     _modelname = "product"
     _identifiers = ("productnumber",)
     _attributes = (
@@ -133,8 +144,9 @@ class Product(DiffSyncModel):
 
     # SEO Fields
     page_title: str = ""
-    meta_description: str= ""
+    meta_description: str = ""
     meta_keywords: str = ""
+
 
 class CategoryToDevice(DiffSyncModel):
     """
@@ -144,12 +156,14 @@ class CategoryToDevice(DiffSyncModel):
         category_name (str): Name of the category.
         productnumber (str): Identifier of the product.
     """
+
     _modelname = "category_to_device"
     _identifiers = ("category_name", "productnumber")
     _attributes = ()
 
     category_name: str
     productnumber: str
+
 
 class AttributeValueToProduct(DiffSyncModel):
     """
@@ -161,6 +175,7 @@ class AttributeValueToProduct(DiffSyncModel):
         value (str): The attribute value.
         price (int): Price override for this attribute value.
     """
+
     _modelname = "attribute_value_to_product"
     _identifiers = ("productnumber", "attribute", "value")
     _attributes = ("price",)
@@ -168,7 +183,8 @@ class AttributeValueToProduct(DiffSyncModel):
     productnumber: str
     attribute: str
     value: str
-    price: int = 0
+    price: float = 0
+
 
 class ProductPhoto(DiffSyncModel):
     """
@@ -183,8 +199,9 @@ class ProductPhoto(DiffSyncModel):
         alttext (str): Alternate text for the photo.
         is_main (bool): Flag indicating if this is the main photo.
     """
+
     _modelname = "product_photo"
-    _identifiers = ("productnumber","alttext","file_type")
+    _identifiers = ("productnumber", "alttext", "file_type")
     _attributes = ("source",)
 
     file_type: str
